requirement.txt
aiosqlite
beautifulsoup4
duckduckgo_search
tavily-python
orjson==3.10.6
langgraph==0.1.15
langsmith
langchain-google-genai
langchain-openai
langchain-community
pillow
fastapi
uvicorn
sqlalchemy
python-dotenv
pydantic>=2.7.0
pydantic-settings
gradio